move interfaces and edit imports for search
