+++ src/components/api/const/interfaces/index.tsx
@@ -2,16 +2,11 @@ import { SingleVenueResponse, SingleVenueResponseContent, SingleVenueProp, Booki
 import { CreateVenueFormInputs } from "./venueFormInterface";
 import { VenueInterface } from "./singleVenueInterfaces";
 import { bookingTileInterface } from "./venueTileInterface";
-import { SearchFormInputInterface } from "./venuesSearchInterface";
-import { ConvertedSearchDataInterface } from "./convertedSearchInterface";
-import { SearchReturnInterface } from "./searchReturnInterface";
-import { Bookings } from "./searchReturnInterface";
+import { SearchFormInputInterface, ConvertedSearchDataInterface, SearchReturnInterface, Bookings, SearchFormInputFP, SearchFormMainProps } from "./searchInterfaces";
 import { FilterValues } from "./facilityFilterInterface";
 import { UpdateProfileInterface, UpdatedProfileData } from "./profileUpdateInterface";
 import { NavBarProps } from "./navBarInterfaces";
 import { ContactFormInputs } from "./contactFormInterface";
-import { SearchFormInputFP } from "./searchInterface";
-import { SearchFormMainProps } from "./convertedSearchInterface";
 import { BookingGuestsInterface, bookingDate } from "./bookingInterfaces";
 import { ProfileMedia, ProfileBookings, ProfileContent, ProfileBioText, ProfileData, VenueManagerBookings, VenueManagerBookingsData } from "./profileInterfaces";
 import { LoginFormInputs, LoginResponse } from "./loginInterfaces";
