--- src/components/api/const/interfaces/searchInterface.tsx
@@ -1,6 +1,0 @@
-export interface SearchFormInputFP {
-  location: string;
-  dateFrom: string;
-  dateTo: string;
-  guests: number;
-}
