+++ src/components/api/const/interfaces/searchInterfaces.tsx
@@ -1,3 +1,10 @@
+export interface SearchFormInputFP {
+  location: string;
+  dateFrom: string;
+  dateTo: string;
+  guests: number;
+}
+
 interface Location {
   city?: string;
   country?: string;
@@ -33,4 +40,24 @@ export interface SearchReturnInterface {
   guests: number;
   meta: Meta;
   media: Media[];
+}
+
+export interface ConvertedSearchDataInterface {
+  dateFrom?: string | undefined;
+  dateTo?: string | undefined;
+  location?: string;
+  name?: string;
+  guests?: number;
+}
+
+export interface SearchFormMainProps {
+  onSearch: (data: ConvertedSearchDataInterface) => void;
+}
+
+export interface SearchFormInputInterface {
+  location?: string;
+  name?: string;
+  dateFrom?: Date;
+  dateTo?: Date;
+  guests?: number;
 }
