--- src/components/api/const/interfaces/venuesSearchInterface.tsx
@@ -1,7 +1,0 @@
-export interface SearchFormInputInterface {
-  location?: string;
-  name?: string;
-  dateFrom?: Date;
-  dateTo?: Date;
-  guests?: number;
-}
